fix up the tests
edges go from parents to children in GraphPoints.
Points with same id and location get merged, so identical swcs do not
produce multiple copies of the points

+++ neurolight/gunpowder/mouselight_swc_file_source.py
@@ -97,7 +97,7 @@ class MouselightSwcFileSource(SwcFileSource):
                             / spacing
                         ).take(self.transpose)
                         * self.scale,
-                        dtype=int,
+                        dtype=float,
                     ),
                     radius=float(row[5]),
                 )
